Inizializzazione GUI SelezionaFiltri

+++ Programma/src/DAO/ImpiegatoDAO.java
@@ -11,12 +11,12 @@ public class ImpiegatoDAO {
 	
 	
 	
-	public void add_Impiegato_To_DB (String cF, String nome, String cognome, String mail, String telefono, float salario, Database DB) {
+	public void add_Impiegato_To_DB (String cF, String nome, String cognome, String mail, String telefono, double salario, Database DB) {
 		Impiegato p1 = new Impiegato(cF, nome, cognome, mail, telefono, salario);
 		DB.add_Impiegato(p1);
 	}
 	
-	public void add_Impiegato_To_DB(String cF, String nome, String cognome, String mail, String telefono, float salario, ArrayList<Skill> skills, Database DB) {
+	public void add_Impiegato_To_DB(String cF, String nome, String cognome, String mail, String telefono, double salario, ArrayList<Skill> skills, Database DB) {
 		Impiegato p1 = new Impiegato(cF, nome, cognome, mail, telefono, salario);
 		p1.addSkill(skills);
 		DB.add_Impiegato(p1);
